Reworked the entirety of Dynamic Auto
I plan on making this a plugin, so I had to make the dynamic a little more... well, dynamic. I made it easier to use, easier to understand, and able to be used on any robot, command based or not. This is the biggest update I've ever done, and probably ever will do. Haven't tested, but needed to get this pushed before I went insane.