fixing for state
I fixed the intake cause the buttons were switched. I put the arm up power to 100% and the arm down power to 40%. I also changed the intake for when the arm goes up cause it was going the wrong direction. Lastly i changed the climb from right joy press to start.